Project Complete; Ready for Review
All in one day, I was successfully able to both start and finish this project. Even with the addition of a secondary HTML page with a corresponding javaScript file as well, I still managed to get the forms within to be working with all the documentation necessary to boot. You can now say that not only this project, but the stack as a whole is completed (and ready for review, of course).